+++ semcore/checkbox/src/Checkbox.jsx
@@ -62,6 +62,8 @@ class CheckboxRoot extends Component {
       hoistDisabled: this.hoistDisabled,
       rootDisabled: this.props.disabled,
       ['aria-label']: this.asProps['aria-label'],
+      ['aria-labelledby']: this.asProps['aria-labelledby'],
+      ['aria-describedby']: this.asProps['aria-describedby'],
     };
   }
 
@@ -79,6 +81,8 @@ class CheckboxRoot extends Component {
           'checkedDefault',
           'label',
           'aria-label',
+          'aria-labelledby',
+          'aria-describedby',
         ]}
       >
         {hasChildren ? (
@@ -95,8 +99,16 @@ class CheckboxRoot extends Component {
 }
 
 class ValueRoot extends Component {
-  static defaultProps = {
-    includeInputProps: inputProps,
+  static defaultProps = (props) => {
+    return {
+      includeInputProps: [
+        ...inputProps,
+        ...(props.includeInputProps ?? []),
+        'aria-label',
+        'aria-labelledby',
+        'aria-describedby',
+      ],
+    };
   };
   static enhance = [autoFocusEnhance(), resolveColorEnhance()];
   static displayName = 'Value';